+++ compiler-plugin/src/main/java/io/ballerina/stdlib/http/compiler/HttpCompilerPlugin.java
@@ -22,6 +22,8 @@ import io.ballerina.projects.plugins.CompilerPlugin;
 import io.ballerina.projects.plugins.CompilerPluginContext;
 import io.ballerina.projects.plugins.codeaction.CodeAction;
 import io.ballerina.stdlib.http.compiler.codeaction.AddHeaderParameterCodeAction;
+import io.ballerina.stdlib.http.compiler.codeaction.AddInterceptorRemoteMethodCodeAction;
+import io.ballerina.stdlib.http.compiler.codeaction.AddInterceptorResourceMethodCodeAction;
 import io.ballerina.stdlib.http.compiler.codeaction.AddPayloadParameterCodeAction;
 import io.ballerina.stdlib.http.compiler.codeaction.AddResponseCacheConfigCodeAction;
 import io.ballerina.stdlib.http.compiler.codeaction.AddResponseContentTypeCodeAction;
@@ -50,7 +52,9 @@ public class HttpCompilerPlugin extends CompilerPlugin {
                 new AddPayloadParameterCodeAction(),
                 new AddHeaderParameterCodeAction(),
                 new AddResponseContentTypeCodeAction(),
-                new AddResponseCacheConfigCodeAction()
+                new AddResponseCacheConfigCodeAction(),
+                new AddInterceptorResourceMethodCodeAction(),
+                new AddInterceptorRemoteMethodCodeAction()
         );
     }
 }